requirements.txt
django==5.1
Pillow==10.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2